install.sh update

+++ install.sh
@@ -1,17 +1,21 @@
 #!/bin/bash
 
-tarname="cheats.tar.gz"
-{ # try
-    path="/usr/local/Cellar/navi/$(navi --version)/libexec"
-    echo "navi path is: $path"
+_=$(command -v navi);
+if [ "$?" != "0" ]; then
+    printf -- "You don't seem to have navi installed.\n";
+    printf -- "Get it: https://github.com/denisidoro/navi\n";
+    printf -- "Exiting with code 127...\n";
+    exit 127;
+fi;
 
-    tar czf ${tarname} ./cheats
-    rm -rf "$path/cheats"
-    mv $tarname "$path/$tarname"
+tarname="cheats.tar.gz";
+path="/usr/local/Cellar/navi/$(navi --version)/libexec";
+echo "navi path is: $path";
 
-    cd $path
-    tar xzf $tarname
-    rm -rf "$path/$tarname"
-} || { # catch
-    exit 1
-}
+tar czf ${tarname} ./cheats;
+rm -rf "$path/cheats";
+mv $tarname "$path/$tarname";
+
+cd $path;
+tar xzf $tarname;
+rm -rf "$path/$tarname";
